pyproject.toml
[project]
name = "palm-9000"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "pipecat-ai[google,whisper]>=0.0.80",
    "numpy>=2.2.6",
    "luma-led-matrix>=1.7.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
]

[dependency-groups]
# Dev-only tools and notebook libs (not installed in production when using `--no-dev`).
dev = [
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    "black>=25.1.0",
    "piper-tts[http]>=1.3.0",
    "isort>=6.0.1",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.3",
    "pandas>=2.3.1",
    "langchain[google-genai]>=0.3.26",
    "langchain-huggingface>=0.3.0",
    "langgraph>=0.5.0",
    "openai-whisper>=20250625",
    "pvporcupine>=3.0.5",
    "pvrecorder>=1.2.7",
    "pyttsx3>=2.98",
    "scipy>=1.16.0",
    "sounddevice>=0.5.2",
    "torch>=2.7.1",
    "transformers>=4.53.0",
    "webrtcvad>=2.0.10",
    "google-genai>=1.24.0",
    "pvleopard>=2.0.5",
    "langdetect>=1.0.9",
    "pyaudio>=0.2.14",
    "llama-index-llms-huggingface-api>=0.6.0",
    "llama-index-embeddings-huggingface>=0.6.0",
    "llama-index-vector-stores-chroma>=0.5.0",
    "llama-index-readers-web>=0.5.0",
    "llama-index-tools-google>=0.6.0",
    "gpiozero>=2.0.1",
    "pigpio>=1.78",
    "spidev>=3.7 ; sys_platform == 'linux'",
    "rpi-gpio>=0.7.1 ; sys_platform == 'linux'",
]
